requirements.txt
jieba==0.42.1
markdownify==0.11.6
py2neo==2021.1.5
rasa==3.4.4
rasa_sdk==3.4.0
requests==2.28.2
